inventix-ml/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
prophet
psycopg2-binary
python-dotenv
